# docs/openAPI.yml
openapi: 3.0.3
info:
  title: Movie Reservation System API
  description: API para el sistema de reservación de películas.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Servidor de desarrollo local
paths:
...
  /api/movies/{id}:
    get:
      summary: Obtener Detalles de una Película
      description: Devuelve los detalles de una película específica.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalles de la película
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  title:
                    type: string
                  description:
                    type: string
                  genre:
                    type: string
                  poster:
                    type: string
                  schedules:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                        date:
                          type: string
                        time:
                          type: string
        '404':
          description: Película no encontrada
  /api/admin/movies:
    post:
      summary: Añadir una Nueva Película
      description: Permite a un administrador añadir una nueva película.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                synopsis:
                  type: string
                nationality:
                  type: string
                qualification:
                  type: string
                distributor:
                  type: string
                trailerUrl:
                  type: string
                posterUrl:
                  type: string
                website:
                  type: string
                releaseDate:
                  type: string
                  format: date
                duration:
                  type: integer
                genres:
                  type: array
                  items:
                    format: string
                languages:
                  type: array
                  items:
                    format: string
                actors:
                  type: array
                  items:
                    format: string
      responses:
        '201':
          description: Película creada
        '400':
          description: Solicitud incorrecta
  /api/admin/movies/{id}:
    put:
      summary: Actualizar una Película
      description: Permite a un administrador actualizar los detalles de una película.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
                synopsis:
                  type: string
                nationality:
                  type: string
                qualification:
                  type: string
                distributor:
                  type: string
                trailerUrl:
                  type: string
                posterUrl:
                  type: string
                website:
                  type: string
                releaseDate:
                  type: string
                  format: date
                duration:
                  type: integer
                genres:
                  type: array
                  items:
                    format: string
                languages:
                  type: array
                  items:
                    format: string
                actors:
                  type: array
                  items:
                    format: string
      responses:
        '200':
          description: Película actualizada
        '400':
          description: Solicitud incorrecta
        '404':
          description: Película no encontrada

    delete:
      summary: Eliminar una Película
      description: Permite a un administrador eliminar una película.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Película eliminada exitosamente
        '404':
          description: Película no encontrada
...